updated public turtlenode.io urls
based on `Effective 2018-06-16: Public access to the individual nodes is no longer permitted to better control traffic flow and load balancing. Please use one of the region load balancers or the public.turtlenode.io hostname.`

not sure if this is the right thing to do? just learning and bored and want to contribute 😊